requirements.txt
botocore==1.12.78
Django==1.10.4
django-bootstrap3==11.0.0
django-ckeditor==5.3.0
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-js-asset==1.1.0
django-mptt==0.9.0
django-suit==0.2.25
docutils==0.14
jmespath==0.9.3
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
